Debugged PyCOMPSs interaction; separated Workflow.
